requirements.txt
pillow
pyzbar
pyperclip
tk
